requirements.txt
python-telegram-bot[job-queue]==22.1
python-dotenv==1.0.1
requests==2.32.3
